requirements.txt
aiogram==3.15.0
aiohttp==3.10.11
python-dotenv==1.0.1
requests==2.31.0
telethon==1.40.0 
gspread==6.2.1
google-auth==2.40.3